requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
plyer==2.1.0
requests==2.31.0
psutil==5.9.6
python-multipart==0.0.6
beautifulsoup4==4.12.2
selenium==4.18.1
webdriver-manager==4.0.1 